+++ CHANGELOG.md
@@ -52,6 +52,8 @@
   [#510](https://github.com/Kong/gateway-operator/pull/510)
 - Added command line flags to configure the certificate generator job's images.
   [#516](https://github.com/Kong/gateway-operator/pull/516)
+- Add `KongPluginBinding` reconciler for Konnect Plugins.
+  [#513](https://github.com/Kong/gateway-operator/pull/513)
 
 ### Fixed